cleanup models

+++ client/lib/models/medication.dart
@@ -1,6 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:hive/hive.dart';
-import 'package:meditrax/models/token.dart';
 
 part "medication.freezed.dart";
 part "medication.g.dart";
@@ -16,10 +15,9 @@ class Medication extends HiveObject with _$Medication {
     @HiveField(2) required double dosage,
     @HiveField(3) required String unit,
     @HiveField(4) required String frequency,
-    @HiveField(5) required int inventory,
+    @HiveField(5) required double inventory,
     @HiveField(6) required String userId,
     @HiveField(7) required DateTime createdAt,
     @HiveField(8) required DateTime updatedAt,
   }) = _Medication;
-
 }
